Add files via upload
Download the zip file of the code, open "Login.java" on an editor and enter your credentials. After that, copy the path of "runLogin.vbs" and paste it in Run application in Windows. Make sure you have java in your machine. Voila!! The script will connect you to the server.